Initial restructure done
Basic communication and functions reestablished

+++ src/shared/Items/ItemSpecs.java
@@ -3,5 +3,5 @@ package shared.Items;
 public class ItemSpecs {
     public String name;
     public String description;
-    public double minimumBid;
+    public int minimumBid;
 }
